Copy files and patch to remove symlinks
Copy the files using Build::Copy plugin.

Need to remove the symlinks because ExtUtils::MakeMaker installs the
contents of the files that the symlinks point to which creates duplicate
data on installation.

In the case of libtensorflow, having two symlinks for each library
multiplies the size of the install by three times.